### 2025-04-07 #8020 @editorialbot
[PRE REVIEW]: KeepDelta: A Python Library for Human-Readable Data Differencing
**Submitting author:** @aslan-ng (<a href="http://orcid.org/0009-0004-3387-4502">Aslan Noorghasemi</a>)
**Repository:** https://github.com/aslan-ng/keepdelta
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.4
