+++ paper/paper.bib
@@ -92,7 +92,7 @@
     number =       {11},
     pages =        {2074-2102},
     keywords =     {graphics for simulation, Monte Carlo, simulation design, simulation reporting, simulation studies},
-    doi =          {https://doi.org/10.1002/sim.8086},
+    doi =          {10.1002/sim.8086},
     url =          {https://onlinelibrary.wiley.com/doi/abs/10.1002/sim.8086},
     eprint =       {https://onlinelibrary.wiley.com/doi/pdf/10.1002/sim.8086},
     abstract =     {Simulation studies are computer experiments that involve creating data by pseudo-random sampling. A key strength of simulation studies is the ability to understand the behavior of statistical methods because some “truth” (usually some parameter/s of interest) is known from the process of generating the data. This allows us to consider properties of methods, such as bias. While widely used, simulation studies are often poorly designed, analyzed, and reported. This tutorial outlines the rationale for using simulation studies and offers guidance for design, execution, analysis, reporting, and presentation. In particular, this tutorial provides a structured approach for planning and reporting simulation studies, which involves defining aims, data-generating mechanisms, estimands, methods, and performance measures (“ADEMP”); coherent terminology for simulation studies; guidance on coding simulation studies; a critical discussion of key performance measures and their estimation; guidance on structuring tabular and graphical presentation of results; and new graphical presentations. With a view to describing recent practice, we review 100 articles taken from Volume 34 of Statistics in Medicine, which included at least one simulation study and identify areas for improvement.},
@@ -123,7 +123,7 @@
     pages =        {385-396},
     year =         {2007},
     issn =         {0304-3800},
-    doi =          {https://doi.org/10.1016/j.ecolmodel.2006.11.005},
+    doi =          {10.1016/j.ecolmodel.2006.11.005},
     url =          {https://www.sciencedirect.com/science/article/pii/S0304380006005400},
     author =       {Craig A. Aumann},
     keywords =     {Simulation, Individual-based model, Hierarchy theory, Assessment, Scientific explanation, Emergent property, Complexity},
@@ -153,4 +153,5 @@
     month =        oct,
     year =         2021,
     language =     "en",
+    doi =          "10.1007/s11277-021-08563-4",
 }
