### 2025-04-07 #8020 @aslan-ng
> **Software report:**
>
> ```
> github.com/AlDanial/cloc v 1.98  T=0.04 s (1625.7 files/s, 110426.3 lines/s)
> -------------------------------------------------------------------------------
> Language                     files          blank        comment           code
> -------------------------------------------------------------------------------
> Python                          56            470            658           2709
> YAML                             5             32              5            205
> Markdown                         2             40              0            159
> TeX                              1             11              0            145
> TOML                             1              4              0             31
> SVG                              1              0              1             13
> -------------------------------------------------------------------------------
> SUM:                            66            557            664           3262
> -------------------------------------------------------------------------------
> ```
>
> **Commit count by author**:
>
> ```
>    174	aslan-ng
>      5	Chris McComb
>      2	Aslan Noorghasemi
> ```

I’ve updated the commit history so that both ‘aslan-ng’ and ‘Aslan Noorghasemi’ now correctly appear under a single identity: Aslan Noorghasemi [anoorghasemi@gmail.com](mailto:anoorghasemi@gmail.com). The commit count has been unified accordingly.
