### 2025-04-07 #8020 @aslan-ng
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> ✅ OK DOIs
>
> - 10.17487/RFC3229 is OK
> - 10.1038/s41586-020-2649-2 is OK
> - 10.1109/EMBC.2018.8513277 is OK
>
> 🟡 SKIP DOIs
>
> - No DOI given, and none found for title: Naive differences of executable code
> - No DOI given, and none found for title: DeepDiff
> - No DOI given, and none found for title: xdelta3
> - No DOI given, and none found for title: xdelta
> - No DOI given, and none found for title: detools
> - No DOI given, and none found for title: Design and Analysis of Simulation Experiments
>
> ❌ MISSING DOIs
>
> - 10.1007/s11277-021-08563-4 may be a valid DOI for title: Data traffic management based on compression and M...
>
> ❌ INVALID DOIs
>
> - https://doi.org/10.1002/sim.8086 is INVALID because of 'https://doi.org/' prefix
> - https://doi.org/10.1016/j.ecolmodel.2006.11.005 is INVALID because of 'https://doi.org/' prefix
> ```

I’ve corrected the DOIs in the references. They no longer include the ‘https://doi.org/’ prefix, and the missing DOI has also been added.
