### 2025-04-17 #8075 @editorialbot
[REVIEW]: KeepDelta: A Python Library for Human-Readable Data Differencing
**Submitting author:** @aslan-ng (<a href="http://orcid.org/0009-0004-3387-4502">Aslan Noorghasemi</a>)
**Repository:** https://github.com/aslan-ng/keepdelta
**Branch with paper.md** (empty if default branch):
**Version:** 0.1.9
