### 2025-04-18 #8075 @tushardave26
👋 Hi, @aslan-ng

First off, congratulations on building such a thoughtful and elegant library! I enjoyed reviewing `KeepDelta`. It's clear that much care went into designing technically sound and user-friendly tool. The focus on human-readable output, native Python data structures, and zero dependencies makes it a valuable tool for the research and dev community. 🙌

While testing `KeepDelta` in a clean `Python 3.12` environment using `venv`, I encountered the following error on import:
```bash
ModuleNotFoundError: No module named 'keepdelta.types'
```

After reviewing the repo, it looks like the issue is related to packaging configuration; specifically, the `keepdelta/types` submodule isn't included in the installed distribution when using `pip install`.

✅ Suggested Fix (Packaging)
Please update your pyproject.toml to explicitly include all subpackages. You can do this by replacing:
```bash
[tool.setuptools]
packages = ["keepdelta"]
```
with
```bash
[tool.setuptools.packages.find]
include = ["keepdelta", "keepdelta.*"]
```
This ensures that nested modules like `keepdelta.types` are appropriately included in the installation.

Once applied, please try importing the `keepdelta` in a fresh `venv`.

🧪 Suggested Test to Catch Packaging Issues
To prevent this kind of packaging/import error in future versions, consider adding a simple automated test to your tests/ directory:
```bash
# tests/test_import.py

def test_keepdelta_import():
    try:
        import keepdelta
        import keepdelta.types
        import keepdelta.types.collections
    except ImportError as e:
        raise AssertionError(f"Import failed: {e}")
```
This will ensure all critical submodules are correctly included and importable after installation. It's a lightweight addition that could be integrated into CI to catch packaging regressions early.

📘 Documentation Suggestion
It may be helpful to include a slightly more complex usage example (e.g., nested dictionaries/lists) in the documentation to showcase the tool's capabilities beyond the basic case. This would better demonstrate the tool's power in real-world scenarios and help new users get started quickly.

✏️ Minor Typos in the Paper
While reviewing the JOSS submission, I noticed just a couple of minor typos you might want to polish up:

Line 30:
"KeepDelta also applied to sensing" → should be → "KeepDelta also applies to sensing"

Line 46:
"dependecies" → should be → "dependencies"

Thanks again for sharing such a well-crafted tool with the open-source and research community. `KeepDelta` is super practical, refreshingly simple, and fills a real gap, especially for those working in simulation and data-heavy environments. I’m excited to see it out in the wild and wish you all the best with the JOSS publication! 🚀

Warm regards,
Tushar (@tushardave26)
