+++ tests/test_import.py
@@ -12,6 +12,12 @@ class TestImport(unittest.TestCase):
             import keepdelta
             import keepdelta.types
             import keepdelta.types.collections
+            import keepdelta.types.primitives
+            import keepdelta.types.primitives.delta_bool
+            import keepdelta.types.primitives.delta_complex
+            import keepdelta.types.primitives.delta_float
+            import keepdelta.types.primitives.delta_int
+            import keepdelta.types.primitives.delta_str
         except ImportError as e:
             raise AssertionError(f"Import failed: {e}")
         
