### 2025-04-19 a7e5b8e
Added examples for primitive types.
