+++ keepdelta/t.py
@@ -1,9 +1,0 @@
-old = 3+4j
-new = 1+5j
-print(new-old)
-
-import keepdelta as kd
-
-old = 1
-new = 1
-print(kd.create(old, new))
