+++ README.md
@@ -48,8 +48,9 @@ pip install keepdelta
 ## Usage
 There are two core methods corresponding to the creation and application of delta encodings:
 
-1. `create(old, new)`:
+### 1. `create(old, new)`
 The `create` function compares the `old` and `new` variables to generate `delta` that captures the differences between two data structures. It produces a compact data structure containing only these differences, and its high human readability greatly aids debugging during development.
+
 #### Example:
 ```python
 >>> import keepdelta as kd
@@ -77,8 +78,10 @@ The `create` function compares the `old` and `new` variables to generate `delta`
 }
 ```
 
-2. `apply(old, delta)`:
+### 2. `apply(old, delta)`
+
 The `apply` function takes the `old` variable and the `delta`, then applies the `delta` to recreate the updated, `new` variable.
+
 #### Example:
 ```python
 >>> import keepdelta as kd
@@ -105,6 +108,7 @@ The `apply` function takes the `old` variable and the `delta`, then applies the 
     "is_student": False
 }
 ```
+
 For more usage examples, refer to the [`examples`](https://github.com/aslan-ng/KeepDelta/tree/main/examples) folder in the project repository.
 
 ## Surpported Formats
@@ -114,8 +118,10 @@ KeepDelta supports common native Python data structures, ensuring compatibility 
 
 <details>
 <summary>
-Boolean (<code>bool</code>) – e.g., True, False
+    <b>Boolean</b> (<code>bool</code>)
 </summary>
+Since booleans aren’t numeric, the delta is simply the new state (True or False).
+
 #### Example:
 ```python
 >>> import keepdelta as kd
@@ -131,12 +137,14 @@ Boolean (<code>bool</code>) – e.g., True, False
 >>> print(delta)
 True
 ```
+<br>
 </details>
 
 <details>
 <summary>
-String (<code>str</code>) – e.g., "hello", "world"
+    <b>String</b> (<code>str</code>)
 </summary>
+
 #### Example:
 ```python
 >>> import keepdelta as kd
@@ -156,8 +164,9 @@ bye
 
 <details>
 <summary>
-Integer (<code>int</code>) – e.g., 42, -7
+    <b>Integer</b> (<code>int</code>)
 </summary>
+
 #### Example:
 ```python
 >>> import keepdelta as kd
@@ -177,8 +186,9 @@ Integer (<code>int</code>) – e.g., 42, -7
 
 <details>
 <summary>
-Float (<code>float</code>) – e.g., 3.14, -0.001
+    <b>Float</b> (<code>float</code>)
 </summary>
+
 #### Example:
 ```python
 >>> import keepdelta as kd
@@ -198,8 +208,9 @@ Float (<code>float</code>) – e.g., 3.14, -0.001
 
 <details>
 <summary>
-Complex (<code>complex</code>) – e.g., 3+4j, -2j
+    <b>Complex</b> (<code>complex</code>)
 </summary>
+
 #### Example:
 ```python
 >>> import keepdelta as kd
@@ -219,8 +230,9 @@ Complex (<code>complex</code>) – e.g., 3+4j, -2j
 
 <details>
 <summary>
-(<code>None</code>)
+    <b>NoneType</b> (<code>None</code>)
 </summary>
+
 #### Example:
 ```python
 >>> import keepdelta as kd
@@ -239,10 +251,122 @@ None
 </details>
 
 ### Collections:
-    * dict – e.g., {"location": "world", "age": 30}
-    * list – e.g., [1, True, "hello"]
-    * tuple – e.g., (2, {"location": "world"}, 3.14)
-    * set – e.g., {1, 2, "apple"}
+<details>
+<summary>
[…]
