### 2025-04-20 e7111d0
supported data structures part enhanced.
