+++ README.md
@@ -111,16 +111,18 @@ The `apply` function takes the `old` variable and the `delta`, then applies the 
 
 For more usage examples, refer to the [`examples`](https://github.com/aslan-ng/KeepDelta/tree/main/examples) folder in the project repository.
 
-## Surpported Formats
-KeepDelta supports common native Python data structures, ensuring compatibility and flexibility when working with a wide variety of data types. The currently supported structures include:
+## Supported Data Structures
+KeepDelta supports common native Python data structures, ensuring compatibility and flexibility when working with a wide variety of data types. The currently supported structures are listed below. Click any item to see how it’s handled and view a quick example:
+
 
 ### Primitive Types:
-
 <details>
 <summary>
     <b>Boolean</b> (<code>bool</code>)
 </summary>
-Since booleans aren’t numeric, the delta is simply the new state (True or False).
+<br>
+
+Since booleans have only two states, the delta is simply the new state (True or False).
 
 #### Example:
 ```python
@@ -144,6 +146,9 @@ True
 <summary>
     <b>String</b> (<code>str</code>)
 </summary>
+<br>
+
+The delta for strings is simply the new string value.
 
 #### Example:
 ```python
@@ -166,6 +171,9 @@ bye
 <summary>
     <b>Integer</b> (<code>int</code>)
 </summary>
+<br>
+
+For integers, the delta is computed as subtraction of values, yielding the offset to apply during reconstruction.
 
 #### Example:
 ```python
@@ -188,6 +196,9 @@ bye
 <summary>
     <b>Float</b> (<code>float</code>)
 </summary>
+<br>
+
+For floats, the delta is computed as subtraction of values, yielding the offset to apply during reconstruction.
 
 #### Example:
 ```python
@@ -210,6 +221,9 @@ bye
 <summary>
     <b>Complex</b> (<code>complex</code>)
 </summary>
+<br>
+
+For complex numbers, the delta is computed as subtraction of values, yielding the offset to apply during reconstruction.
 
 #### Example:
 ```python
@@ -232,6 +246,9 @@ bye
 <summary>
     <b>NoneType</b> (<code>None</code>)
 </summary>
+<br>
+
+Since KeepDelta supports type change, it is possible to track the changes from `None` to other types or vise versa.
 
 #### Example:
 ```python
@@ -255,6 +272,9 @@ None
 <summary>
     <b>Dictionary</b> (<code>dict</code>)
 </summary>
+<br>
+
+When diffing dictionaries, key-value pairs in the inputs are compared. The key removal is marked with the special token `__delete__`.
 
 #### Example:
 ```python
@@ -262,22 +282,28 @@ None
 
 >>> # Initial data
 >>> old = {
-        "location": "earth",
-        "age": 20
-    }
+...     "location": "earth",
+...     "age": 20,
+...     "snacks": ["chocolate", "bananas"],
+...     "student": True,  # Will be removed.
+... }
 
 >>> # Updated data
 >>> new = {
-        "location": "mars",
-        "age": 30
-    }
+...     "location": "mars",
+...     "age": 30,
+...     "snacks": ["chocolate", "bananas"],
+...     "happy": True,  # The newly added key
+... }
 
 >>> # Create delta
 >>> delta = kd.create(old, new)
 >>> print(delta)
 {
     "location": "mars",
-    "age": 10
+    "age": 10,
+    "student": "__delete__",  # The removed key
+    "happy": True  # The newly added key
 }
 ```
 </details>
@@ -286,8 +312,9 @@ None
 <summary>
     <b>List</b> (<code>list</code>)
 </summary>
+<br>
 
-The delta for a list is a dictionary where each key is a list index and each value describes the change applied at that position; including a numerical offset (to adjust the original element) or __delete__ (to remove it).
[…]
