### 2025-04-20 5b0b859
examples have commented outputs.
