### 2025-04-20 c8aa84c
testing similar and equal types added.
