+++ tests/test_type_conversion/test_type_conversion_tuple.py
@@ -3,26 +3,26 @@ import unittest
 import keepdelta as kd
 
 
-class TestChangeTypeDict(unittest.TestCase):
+class TestTypeConversionTuple(unittest.TestCase):
 
-    def test_none_to_dict(self):
+    def test_none_to_tuple(self):
         old = None
-        new = {
-            "str": "hello",
-            "bool": True,
-        }
-        delta = {
-            "str": "hello",
-            "bool": True,
-        }
+        new = (
+            "hello",
+            True,
+        )
+        delta = (
+            "hello",
+            True,
+        )
         self.assertEqual(kd.create(old, new), delta)
         self.assertEqual(kd.apply(old, delta), new)
 
-    def test_dict_to_none(self):
-        old = {
-            "str": "hello",
-            "bool": True,
-        }
+    def test_tuple_to_none(self):
+        old = (
+            "hello",
+            True,
+        )
         new = None
         delta = None
         self.assertEqual(kd.create(old, new), delta)
