+++ tests/test_type_conversion/test_type_conversion_str.py
@@ -3,33 +3,33 @@ import unittest
 import keepdelta as kd
 
 
-class TestChangeTypeBool(unittest.TestCase):
+class TestTypeConversionStr(unittest.TestCase):
 
-    def test_bool_to_complex(self):
-        old = True
+    def test_str_to_bool(self):
+        old = "hello"
+        new = True
+        delta = True
+        self.assertEqual(kd.create(old, new), delta)
+        self.assertEqual(kd.apply(old, delta), new)
+
+    def test_str_to_complex(self):
+        old = "hello"
         new = 1 + 1j
         delta = 1 + 1j
         self.assertEqual(kd.create(old, new), delta)
         self.assertEqual(kd.apply(old, delta), new)
 
-    def test_bool_to_float(self):
-        old = True
-        new = 10.1
-        delta = 10.1
+    def test_str_to_float(self):
+        old = "hello"
+        new = 1.1
+        delta = 1.1
         self.assertEqual(kd.create(old, new), delta)
         self.assertEqual(kd.apply(old, delta), new)
 
-    def test_bool_to_int(self):
-        old = True
-        new = 10
-        delta = 10
-        self.assertEqual(kd.create(old, new), delta)
-        self.assertEqual(kd.apply(old, delta), new)
-
-    def test_bool_to_str(self):
-        old = True
-        new = "hello"
-        delta = "hello"
+    def test_str_to_int(self):
+        old = "hello"
+        new = 1
+        delta = 1
         self.assertEqual(kd.create(old, new), delta)
         self.assertEqual(kd.apply(old, delta), new)
 
