+++ tests/test_type_conversion/test_type_conversion_float.py
@@ -3,33 +3,33 @@ import unittest
 import keepdelta as kd
 
 
-class TestChangeTypeStr(unittest.TestCase):
+class TestTypeConversionFloat(unittest.TestCase):
 
-    def test_str_to_bool(self):
-        old = "hello"
+    def test_float_to_bool(self):
+        old = 1.1
         new = True
         delta = True
         self.assertEqual(kd.create(old, new), delta)
         self.assertEqual(kd.apply(old, delta), new)
 
-    def test_str_to_complex(self):
-        old = "hello"
+    def test_float_to_complex(self):
+        old = 1.1
         new = 1 + 1j
         delta = 1 + 1j
         self.assertEqual(kd.create(old, new), delta)
         self.assertEqual(kd.apply(old, delta), new)
 
-    def test_str_to_float(self):
-        old = "hello"
-        new = 1.1
-        delta = 1.1
+    def test_float_to_int(self):
+        old = 1.1
+        new = 1
+        delta = 1
         self.assertEqual(kd.create(old, new), delta)
         self.assertEqual(kd.apply(old, delta), new)
 
-    def test_str_to_int(self):
-        old = "hello"
-        new = 1
-        delta = 1
+    def test_float_to_str(self):
+        old = 1.1
+        new = "hello"
+        delta = "hello"
         self.assertEqual(kd.create(old, new), delta)
         self.assertEqual(kd.apply(old, delta), new)
 
