+++ tests/test_type_conversion/test_type_conversion_bool.py
@@ -3,31 +3,31 @@ import unittest
 import keepdelta as kd
 
 
-class TestChangeTypeInt(unittest.TestCase):
+class TestTypeConversionBool(unittest.TestCase):
 
-    def test_int_to_bool(self):
-        old = 1
-        new = True
-        delta = True
-        self.assertEqual(kd.create(old, new), delta)
-        self.assertEqual(kd.apply(old, delta), new)
-
-    def test_int_to_complex(self):
-        old = 1
+    def test_bool_to_complex(self):
+        old = True
         new = 1 + 1j
         delta = 1 + 1j
         self.assertEqual(kd.create(old, new), delta)
         self.assertEqual(kd.apply(old, delta), new)
 
-    def test_int_to_float(self):
-        old = 1
-        new = 1.1
-        delta = 1.1
+    def test_bool_to_float(self):
+        old = True
+        new = 10.1
+        delta = 10.1
         self.assertEqual(kd.create(old, new), delta)
         self.assertEqual(kd.apply(old, delta), new)
 
-    def test_int_to_str(self):
-        old = 1
+    def test_bool_to_int(self):
+        old = True
+        new = 10
+        delta = 10
+        self.assertEqual(kd.create(old, new), delta)
+        self.assertEqual(kd.apply(old, delta), new)
+
+    def test_bool_to_str(self):
+        old = True
         new = "hello"
         delta = "hello"
         self.assertEqual(kd.create(old, new), delta)
