+++ tests/test_type_conversion/test_type_conversion_set.py
@@ -3,26 +3,26 @@ import unittest
 import keepdelta as kd
 
 
-class TestChangeTypeTuple(unittest.TestCase):
+class TestTypeConversionSet(unittest.TestCase):
 
-    def test_none_to_tuple(self):
+    def test_none_to_set(self):
         old = None
-        new = (
+        new = {
             "hello",
             True,
-        )
-        delta = (
+        }
+        delta = {
             "hello",
             True,
-        )
+        }
         self.assertEqual(kd.create(old, new), delta)
         self.assertEqual(kd.apply(old, delta), new)
 
-    def test_tuple_to_none(self):
-        old = (
+    def test_set_to_none(self):
+        old = {
             "hello",
             True,
-        )
+        }
         new = None
         delta = None
         self.assertEqual(kd.create(old, new), delta)
