+++ .github/workflows/run-tests.yml
@@ -7,7 +7,6 @@ on:
     paths:
       - keepdelta/**
       - tests/**
-      - setup.py
       - pyproject.toml
       - .github/workflows/run-tests.yml
   pull_request:
@@ -16,7 +15,6 @@ on:
     paths:
       - keepdelta/**
       - tests/**
-      - setup.py
       - pyproject.toml
       - .github/workflows/run-tests.yml
 
