+++ .github/workflows/run-examples-pypi.yml
@@ -1,19 +1,20 @@
 name: Run Examples PyPI
 
 on:
+  workflow_dispatch:
+
   push:
     branches:
       - main
     paths:
-      - keepdelta/**
       - examples/**
       - pyproject.toml
       - .github/workflows/run-examples-pypi.yml
+      
   pull_request:
     branches:
       - main
     paths:
-      - keepdelta/**
       - examples/**
       - pyproject.toml
       - .github/workflows/run-examples-pypi.yml
