+++ .github/workflows/run-tests.yml
@@ -9,6 +9,7 @@ on:
       - tests/**
       - pyproject.toml
       - .github/workflows/run-tests.yml
+
   pull_request:
     branches:
       - main
