### 2025-04-20 #8075 @aslan-ng
> 👋 Hi, [@aslan-ng](https://github.com/aslan-ng)
>
> First off, congratulations on building such a thoughtful and elegant library! I enjoyed reviewing `KeepDelta`. It's clear that much care went into designing technically sound and user-friendly tool. The focus on human-readable output, native Python data structures, and zero dependencies makes it a valuable tool for the research and dev community. 🙌
>
> While testing `KeepDelta` in a clean `Python 3.12` environment using `venv`, I encountered the following error on import:
>
> ModuleNotFoundError: No module named 'keepdelta.types'
> After reviewing the repo, it looks like the issue is related to packaging configuration; specifically, the `keepdelta/types` submodule isn't included in the installed distribution when using `pip install`.
>
> ✅ Suggested Fix (Packaging) Please update your pyproject.toml to explicitly include all subpackages. You can do this by replacing:
>
> [tool.setuptools]
> packages = ["keepdelta"]
> with
>
> [tool.setuptools.packages.find]
> include = ["keepdelta", "keepdelta.*"]
> This ensures that nested modules like `keepdelta.types` are appropriately included in the installation.
>
> Once applied, please try importing the `keepdelta` in a fresh `venv`.
>
> 🧪 Suggested Test to Catch Packaging Issues To prevent this kind of packaging/import error in future versions, consider adding a simple automated test to your tests/ directory:
>
> # tests/test_import.py
>
> def test_keepdelta_import():
>     try:
>         import keepdelta
>         import keepdelta.types
>         import keepdelta.types.collections
>     except ImportError as e:
>         raise AssertionError(f"Import failed: {e}")
> This will ensure all critical submodules are correctly included and importable after installation. It's a lightweight addition that could be integrated into CI to catch packaging regressions early.
>
> 📘 Documentation Suggestion It may be helpful to include a slightly more complex usage example (e.g., nested dictionaries/lists) in the documentation to showcase the tool's capabilities beyond the basic case. This would better demonstrate the tool's power in real-world scenarios and help new users get started quickly.
>
> ✏️ Minor Typos in the Paper While reviewing the JOSS submission, I noticed just a couple of minor typos you might want to polish up:
>
> Line 30: "KeepDelta also applied to sensing" → should be → "KeepDelta also applies to sensing"
>
> Line 46: "dependecies" → should be → "dependencies"
>
> Thanks again for sharing such a well-crafted tool with the open-source and research community. `KeepDelta` is super practical, refreshingly simple, and fills a real gap, especially for those working in simulation and data-heavy environments. I’m excited to see it out in the wild and wish you all the best with the JOSS publication! 🚀
>
> Warm regards, Tushar ([@tushardave26](https://github.com/tushardave26))

Dear @tushardave26 ,

Thank you so much for the kind words and thoughtful review; it truly means a lot! I’m really glad you found KeepDelta useful and appreciated the design choices. I hope it proves helpful to the broader community as well.

✅ I’ve addressed all the issues you raised:
- Updated the packaging configuration in pyproject.toml to properly include all subpackages (including keepdelta.types)
- Added the suggested import test to catch future packaging regressions in CI, plus github actions to run tests and examples using latest pypi releases.
- Expanded the README with more complex examples (nested dictionaries, sets, and mixed types) to better illustrate KeepDelta’s real-world behavior. They are available as collapsible under [Supported Data Types & Behaviors](https://github.com/aslan-ng/keepdelta/tree/v0.1.5#supported-data-types--behaviors) section.
- Corrected the typos in the JOSS paper. Thanks for spotting those!

After making these changes, I also bumped the version to v0.1.5 and published the updated package to PyPI.

Thanks again for your time and constructive feedback. I really appreciate your help in improving KeepDelta!

Warmly,
Aslan
