### 2025-04-20 76aa5c2
added newlines for readibility to the end of examples.
