### 2025-04-20 b52f1bb
better example for float
