### 2025-04-20 #8075 @tushardave26
@crvernon - The author has addressed all the issues I raised.

The packaging issue has been fixed, the documentation is more precise, and the additional example guidance adds valuable context for new users. The tool functions as described and remains a lightweight, user-friendly, and well-scoped contribution to the Python ecosystem, especially for researchers working with dynamic data structures.

I genuinely appreciate the authors’ responsiveness and the care put into refining the software and the manuscript.

 I have also gone through my checklist. I’m happy to recommend this submission for publication in JOSS.
