+++ .github/workflows/run-examples-pypi.yml
@@ -3,6 +3,9 @@ name: Run Examples PyPI
 on:
   workflow_dispatch:
 
+  release:
+    types: [published]
+
   push:
     branches:
       - main
