+++ .github/workflows/pypi-release.yml
@@ -7,6 +7,10 @@ on:
 jobs:
   deploy:
     runs-on: ubuntu-latest
+    environment: pypi
+    permissions:
+      id-token: write
+      contents: read
 
     steps:
       - name: Checkout repository
@@ -20,26 +24,10 @@ jobs:
       - name: Install build tools
         run: |
           python -m pip install --upgrade pip
-          python -m pip install build twine
+          python -m pip install build
 
       - name: Build package
         run: python -m build
 
       - name: Publish to PyPI
-        env:
-          TWINE_USERNAME: __token__
-          TWINE_PASSWORD: ${{ secrets.PYPI_API_TOKEN }}
-        run: twine upload --repository pypi dist/*
-
-      - name: Install KeepDelta from PyPI
-        run: pip install keepdelta
-
-      - name: Run unit tests
-        run: python tests/tests.py
-
-      - name: Run all example scripts
-        run: |
-          for script in examples/*.py; do
-            echo "Running $script"
-            python "$script"
-          done
+        uses: pypa/gh-action-pypi-publish@release/v1
