+++ pyproject.toml
@@ -13,7 +13,7 @@ keepdelta = ["py.typed"]
 
 [project]
 name = "keepdelta"
-version = "0.1.5"
+version = "0.1.6"
 description = "Efficient Data Differencing for Python Data Structures"
 readme = "README.md"
 license = {text = "MIT"}
