### 2025-04-21 ad6c828
removed urls since it is not required anymore
