+++ pyproject.toml
@@ -31,7 +31,3 @@ classifiers = [
     "Topic :: Scientific/Engineering :: Information Analysis",
     "Typing :: Typed",
 ]
-
-[project.urls]
-Homepage = "https://github.com/aslan-ng/keepdelta"
-Issues = "https://github.com/aslan-ng/keepdelta/issues"
