### 2025-05-08 #8075 @ujjwalkarn
👋 Hi @aslan-ng, within the "Comparison to Existing Tools" section in the paper:

1. for "deepdiff", is it possible to add a line comparing its features and optimization options with KeepDelta? My understanding is that deepdiff is pretty comprehensive and has some configuration options, so would be great to see how it compares with KeepDelta in this regard.
2. another library "datadiff" (https://pypi.org/project/datadiff/) claims to provide human-readable outputs too, could we add it as a comparison?
3. "difflib" seems to be a part of the Python Standard Library requiring no external dependencies, could we add a line about this too?
