### 2025-05-08 #8075 @ujjwalkarn
@aslan-ng, in [README.md](https://github.com/aslan-ng/keepdelta/blob/main/README.md), the figure under "What is Delta Encoding?", both the full-state encoding (left) vs. delta encoding (right) have the same diagram in different colors. Is it possible to update the delta encoding diagram make the distinction clearer? (maybe somehow showing that it uses "deltas" or differences?)
