### 2025-05-08 a5b1d06
last paragraph has less repetitive wording
