+++ paper/paper.md
@@ -53,6 +53,6 @@ Human-readability is crucial for debugging and development in both simulation an
 In the landscape of Python libraries designed for delta encoding, several notable tools have emerged, each with distinct features and applications.
 `xdelta3` [@xdelta3] and its predecessor, `xdelta` [@xdelta], are tools that perform delta encoding at the binary level. These utilities are particularly effective for binary file differencing and are widely used in version control systems and data synchronization tasks. However, both are considered outdated and are no longer actively maintained. Their operation at the binary level results in outputs that are not human-readable, and they are not tailored for Python data structures, limiting their applicability in Python-centric workflows.
 `deepdiff` [@deepdiff] is a contemporary library that facilitates the identification of differences between complex Python data structures, including dictionaries, lists, and sets. It extends support to external libraries like `NumPy` [@numpy], enhancing its versatility. However, this integration can lead to outputs that are less human-readable compared to `KeepDelta`. Additionally, deepdiff introduces weighty dependencies that may not be necessary for all projects.
-In contrast to these alternatives, `KeepDelta` is a lightweight Python library optimized for simulations, focusing on efficient delta management for native Python data structures. It operates directly on Python’s native data types, producing human-readable outputs that facilitate debugging and research applications. Implemented in pure Python, `KeepDelta` eliminates external dependencies, making it an ideal choice for Python-centric workflows, particularly in simulations and data analysis tasks.
+In contrast to these alternatives, `KeepDelta` is a lightweight Python library tailored for simulations and sensing, focusing on efficient delta management for built-in Python data structures. It produces human-readable outputs that facilitate debugging and research workflows. Written entirely in Python and free of external dependencies, `KeepDelta` integrates seamlessly into Python-centric workflows, including simulation, sensing, and data analysis pipelines.
 
 # References
