### 2025-05-08 e9979ce
Summary has definition of full-state encoding
