### 2025-05-10 3a96547
figure color changed
