+++ paper/paper.bib
@@ -51,6 +51,14 @@
     year =         {2016},
 }
 
+@software{datadiff,
+    author =       {Brondsema, Dave and R. Coombs, Jason},
+    title =        {DataDiff},
+    year =         {2023},
+    version =      {2.2.0},
+    url =          {https://pypi.org/project/datadiff/},
+}
+
 @software{detools,
     author =       {Moqvist, Erik},
     title =        {{detools}},
