### 2025-05-10 2fffb61
smart home example added instead of previous student profile
