### 2025-05-10 1691450
example image and captioninserted into paper.
