+++ paper/paper.md
@@ -48,16 +48,22 @@ Both simulations and sensing require mechanisms to track and store evolving stat
 
 Human-readability is crucial for debugging and development in both simulation and sensing projects, and `KeepDelta` is tailored specifically for this purpose. It is lightweight and has no dependencies, supports native Python data structures, and generates results that are easy to interpret. This makes it an ideal tool for Python developers and researchers seeking a simple yet powerful solution for change management. By providing clear, human-readable output, `KeepDelta` enhances both the development process and debugging efficiency, making it easier to track and manage changes in complex projects involving both computational models and real-world sensing systems.
 
-![An example of human-readable change tracking with KeepDelta. (a) Previous smart-home state. (b) State after updates to temperature, lights, door lock, schedule, and alerts. (c) Delta produced by KeepDelta, which records only the changes—numeric adjustments, boolean toggles, a tuple insertion, and a new set element, offering a concise, easily interpretable summary.](./assets/example.png)
+![An example of human-readable change tracking with KeepDelta. (a) Previous smart-home state. (b) State after updates to temperature, lights, door lock, schedule, and alerts. (c) Delta produced by KeepDelta, which records only the changes including numeric adjustments, boolean toggles, a tuple insertion, and a new set element, offering a concise, easily interpretable summary.](./assets/example.png)
 
 # Comparison to Existing Tools
 
 In the landscape of Python libraries designed for delta encoding, several notable tools have emerged, each with distinct features and applications.
+
 `xdelta3` [@xdelta3] and its predecessor, `xdelta` [@xdelta], are tools that perform delta encoding at the binary level. These utilities are particularly effective for binary file differencing and are widely used in version control systems and data synchronization tasks. However, both are considered outdated and are no longer actively maintained. Their operation at the binary level results in outputs that are not human-readable, and they are not tailored for Python data structures, limiting their applicability in Python-centric workflows.
+
 `detools` [@detools] is a Python package that focuses on applying and generating binary patches using a custom delta algorithm optimized for embedded systems. While it is efficient and actively maintained, `detools` operates strictly at the binary level, producing outputs that are not human-readable and lacking support for native Python data structures. As such, it is more suitable for firmware and embedded development than for Python-centric simulation or data analysis workflows.
+
 `difflib` is part of the Python Standard Library and requires no external dependencies. It provides unified- and context‑style text diffs for strings and lists, making it readily available but limited to sequence comparison rather than structured data differencing. It does not expose a programmatic delta object, so you cannot save its diff output as a delta for later state reconstruction.
+
 `datadiff` [@datadiff] offers human-readable, unified-diff style comparisons of Python objects, including nested dictionaries, sequences, sets, and multi-line strings, and integrates with testing frameworks via nose-assert. Similar to `difflib`, it produces text-based diffs, but unlike `difflib`, it handles arbitrary Python objects rather than only strings or lists. However, like `difflib`, `datadiff` only generates one-off diffs and does not expose a programmatic delta object for incremental change tracking or later state reconstruction.
+
 `deepdiff` [@deepdiff] is a contemporary library that facilitates the identification of differences between complex Python data structures, including dictionaries, lists, and sets. It extends support to external libraries like `NumPy` [@numpy], enhancing its versatility. DeepDiff also offers a range of configuration options, such as ignoring iterable order, controlling significant‐digit precision in numeric comparisons, custom comparator functions, and an optional `deepdiff[optimize]` installation for performance tuning in certain processes such as json serialization which makes it highly flexible but adds complexity and dependencies. Also, this extensibility can lead to outputs that are less human-readable compared to KeepDelta, and the added dependencies may not be necessary for all projects.
+
[…]
