### 2025-05-10 2c7cf59
minor enhancements to example
