### 2025-05-11 ae935e0
minor enhancements in tool comparison section
