+++ paper/paper.md
@@ -60,9 +60,9 @@ In the landscape of Python libraries designed for delta encoding, several notabl
 
 `difflib` is part of the Python Standard Library and requires no external dependencies. It provides unified- and context‑style text diffs for strings and lists, making it readily available but limited to sequence comparison rather than structured data differencing. It does not expose a programmatic delta object, so you cannot save its diff output as a delta for later state reconstruction.
 
-`datadiff` [@datadiff] offers human-readable, unified-diff style comparisons of Python objects, including nested dictionaries, sequences, sets, and multi-line strings, and integrates with testing frameworks via nose-assert. Similar to `difflib`, it produces text-based diffs, but unlike `difflib`, it handles arbitrary Python objects rather than only strings or lists. However, like `difflib`, `datadiff` only generates one-off diffs and does not expose a programmatic delta object for incremental change tracking or later state reconstruction.
+`DataDiff` [@datadiff] offers human-readable, unified-diff style comparisons of Python objects, including nested dictionaries, sequences, sets, and multi-line strings, and integrates with testing frameworks via nose-assert. Similar to `difflib`, it produces text-based diffs, but unlike `difflib`, it handles arbitrary Python objects rather than only strings or lists. However, like `difflib`, `DataDiff` only generates one-off diffs and does not expose a programmatic delta object for incremental change tracking or later state reconstruction.
 
-`deepdiff` [@deepdiff] is a contemporary library that facilitates the identification of differences between complex Python data structures, including dictionaries, lists, and sets. It extends support to external libraries like `NumPy` [@numpy], enhancing its versatility. DeepDiff also offers a range of configuration options, such as ignoring iterable order, controlling significant‐digit precision in numeric comparisons, custom comparator functions, and an optional `deepdiff[optimize]` installation for performance tuning in certain processes such as json serialization which makes it highly flexible but adds complexity and dependencies. Also, this extensibility can lead to outputs that are less human-readable compared to KeepDelta, and the added dependencies may not be necessary for all projects.
+`DeepDiff` [@deepdiff] is a contemporary library that facilitates the identification of differences between complex Python data structures, including dictionaries, lists, and sets. It extends support to external libraries like `NumPy` [@numpy], enhancing its versatility. `DeepDiff` also offers a range of configuration options, such as ignoring iterable order, controlling significant‐digit precision in numeric comparisons, and custom comparator functions which make it highly flexible but add complexity and dependencies. Moreover, this extensibility can lead to outputs that are less human-readable compared to `KeepDelta`, and the added dependencies may not be necessary for all projects.
 
 In contrast to these alternatives, `KeepDelta` is a lightweight Python library tailored for simulations and sensing, focusing on efficient delta management for built-in Python data structures. It produces human-readable outputs that facilitate debugging and research workflows. Written entirely in Python and free of external dependencies, `KeepDelta` integrates seamlessly into Python-centric workflows, including simulation, sensing, and data analysis pipelines.
 
