### 2025-05-11 #8075 @aslan-ng
> 👋 Hi [@aslan-ng](https://github.com/aslan-ng), within the "Comparison to Existing Tools" section in the paper:
>
> 1. for "deepdiff", is it possible to add a line comparing its features and optimization options with KeepDelta? My understanding is that deepdiff is pretty comprehensive and has some configuration options, so would be great to see how it compares with KeepDelta in this regard.
> 2. another library "datadiff" (https://pypi.org/project/datadiff/) claims to provide human-readable outputs too, could we add it as a comparison?
> 3. "difflib" seems to be a part of the Python Standard Library requiring no external dependencies, could we add a line about this too?

Hi there!
First of all, thanks for the clear, constructive suggestions!

I’ve updated the “Comparison to Existing Tools” section accordingly:
1. `DeepDiff`: I added a few sentences summarizing its key configuration features such as ignore order, significant-digit precision, and custom comparators, and contrasted those with KeepDelta’s lightweight, dependency-free and fool-proof design.
2. `DataDiff`: I inserted a new paragraph explaining its human-readable unified-diff output, its similarities to difflib, and why it isn’t suited for incremental state reconstruction like KeepDelta.
3. `difflib`: added a brief paragraph noting its zero-dependency status, its focus on sequence text diffs, and the absence of a programmatic delta object.

Please let me know if you’d like any further tweaks!
