### 2025-05-11 #8075 @aslan-ng
> [@aslan-ng](https://github.com/aslan-ng), in [README.md](https://github.com/aslan-ng/keepdelta/blob/main/README.md), the figure under "What is Delta Encoding?", both the full-state encoding (left) vs. delta encoding (right) have the same diagram in different colors. Is it possible to update the delta encoding diagram make the distinction clearer? (maybe somehow showing that it uses "deltas" or differences?)
>
> <img alt="Image" width="642" src="https://private-user-images.githubusercontent.com/5948390/441723081-1552b0cb-94d1-414e-9014-b3dab83696a5.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3NDY5MjM1MTUsIm5iZiI6MTc0NjkyMzIxNSwicGF0aCI6Ii81OTQ4MzkwLzQ0MTcyMzA4MS0xNTUyYjBjYi05NGQxLTQxNGUtOTAxNC1iM2RhYjgzNjk2YTUucG5nP1gtQW16LUFsZ29yaXRobT1BV1M0LUhNQUMtU0hBMjU2JlgtQW16LUNyZWRlbnRpYWw9QUtJQVZDT0RZTFNBNTNQUUs0WkElMkYyMDI1MDUxMSUyRnVzLWVhc3QtMSUyRnMzJTJGYXdzNF9yZXF1ZXN0JlgtQW16LURhdGU9MjAyNTA1MTFUMDAyNjU1WiZYLUFtei1FeHBpcmVzPTMwMCZYLUFtei1TaWduYXR1cmU9ZmVlODgzNDliNzBjNThiNjcxYmY5OTVjMjI5ZmVmZTFlODllNmMzNzk4YzJhNjFiM2IwNTFlYjhhZTc0MzJkZCZYLUFtei1TaWduZWRIZWFkZXJzPWhvc3QifQ.4BcGNJD2W2Z-9EZd6xBSp8fCjk-QLiFFzRfUtSYlkZI">

Dear [ujjwalkarn](https://github.com/ujjwalkarn),

I made some changes to the figure to make it clearer for the user to distinguish between the two approaches. The new version is a clear improvement with those orange Δ-labeled arrows immediately signaling “differences” and making the delta path visually distinct from the full-state path. I also rewrote the caption to call out exactly what the arrows represent (deltas) and how this differs from the full-state approach.

Please let me know if it still needs any tweaks or additional clarity.

Thanks again for the helpful suggestion!
