### 2025-05-11 #8075 @aslan-ng
> All three of these lines seem to refer KeepDelta as being written in native Python ("native Python data structures", "Python’s native data types", "pure Python"), which seems a bit repetitive, maybe these lines can be reworded for better readability.
>
> A couple of other points:
>
> 1. In the paper, is it possible to provide an example of how the outputs from KeepDelta are "human readable"?
> 2. In the paper, "Full-state encoding" isn't formally defined. We have this line "In simulations, the naive approach of saving full snapshots at every timestep leads to excessive storage demands...", can we clarify if this is what "Full-state encoding" refers to?

You’re absolutely right about the repetition around those words. I’ve cleaned up those phrases for better readability.

1. I added Figure 3 to provide a concrete example of KeepDelta’s human-readable output. I think it’s a helpful addition that gives readers a more tangible sense of how the library represents changes concisely.

2. In the Summary section, I clarified what “full-state encoding” refers to. It now reads: "In simulations, the traditional method known as full-state encoding stores the entire system state, including all nested data structures and variable values, at every timestep. While simple to implement, this approach is highly storage-intensive."

Please let me know if there’s anything else that could be clarified or improved!
