### 2025-05-11 a8c4fb5
logo size increased
