+++ README.md
@@ -26,6 +26,7 @@
 *KeepDelta* is a lightweight Python library designed to efficiently track and manage changes (deltas) between Python built-in types. It is applicable to various scenarios  that require dynamic data management, especially when incremental numerical changes are present, such as simulations and sensing. While many alternative tools operate at the binary level, KeepDelta emphasizes human-readable delta encoding, facilitating debugging and analysis for Python developers and researchers across multiple domains.
 
 ## What is Delta Encoding?
+
 In many computational scenarios, efficiently managing evolving data is crucial. Traditional methods, that rely on full-state encoding — which means storing and/or transmitting complete snapshots at each step — can be inefficient due to the large size of the snapshots. Delta encoding addresses this challenge by capturing and applying only the changes (deltas) between successive states of data structures, resulting in significantly smaller and more manageable data.
 
 <div align="center">
@@ -39,6 +40,7 @@ In many computational scenarios, efficiently managing evolving data is crucial. 
 </div>
 
 ## Features
+
 * Generates compact and human-readable differences between two Python variables.
 * Applies delta to a variable to reconstruct the updated version.
 * Supports common Python built-in data types.
@@ -46,18 +48,22 @@ In many computational scenarios, efficiently managing evolving data is crucial. 
 * Requires no external dependencies.
 
 ## Installation
+
 Install the package using pip:
 ```sh
 pip install keepdelta
 ```
 
 ## Usage
+
 There are two core methods corresponding to the creation and application of delta encodings:
 
 ### 1. `create(old, new)`:
+
 The `create` function compares the `old` and `new` variables to generate `delta` that captures the differences between two data structures. It produces a compact data structure containing only these differences, and its high human readability greatly aids debugging during development.
 
 #### Example:
+
 ```python
 >>> import keepdelta as kd
 
@@ -89,6 +95,7 @@ The `create` function compares the `old` and `new` variables to generate `delta`
 The `apply` function takes the `old` variable and the `delta`, then applies the `delta` to recreate the updated, `new` variable.
 
 #### Example:
+
 ```python
 >>> import keepdelta as kd
 
@@ -117,12 +124,12 @@ The `apply` function takes the `old` variable and the `delta`, then applies the 
 
 For more usage examples, refer to the [`examples`](https://github.com/aslan-ng/KeepDelta/tree/main/examples) folder in the project repository, or continue to the next section, "[Supported Data Types & Behaviors](#supported-data-types--behaviors)”, for a detailed look at how each structure is handled.
 
-
 ## Supported Data Types & Behaviors
+
 KeepDelta supports common native Python data structures, ensuring compatibility and flexibility when working with a wide variety of data types. The currently supported structures are listed below. Click any item to see how it’s handled and view a quick example:
 
-
 ### 🔸 Primitive Types:
+
 <details>
 <summary>
     <b>Boolean</b> (<code>bool</code>)
@@ -132,6 +139,7 @@ KeepDelta supports common native Python data structures, ensuring compatibility 
 Since booleans have only two states, the delta is simply the new state (True or False).
 
 #### Example:
+
 ```python
 >>> import keepdelta as kd
 
@@ -158,6 +166,7 @@ True
 The delta for strings is simply the new string value.
 
 #### Example:
+
 ```python
 >>> import keepdelta as kd
 
@@ -184,6 +193,7 @@ bye
 For integers, the delta is computed as subtraction of values, yielding the offset to apply during reconstruction.
 
 #### Example:
+
 ```python
 >>> import keepdelta as kd
[…]
