### 2025-05-30 #8075 @crvernon
👋  @aslan-ng - This submission is looking great.  Please see to the following:

- [ ] Check out the potential missing DOI fix listed in the above comment.
- [ ] LINE 127 (and the in-text citation):  Please leave the draft paper citation out of the paper since the content in this link will change and it seems to be already broken.

So here is what we have left to do:

- [ ]  Conduct a GitHub release of the current reviewed version of the software and archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository) if you have not done so already.  **Please ensure that the software archive uses the same license as the license you have posted on code repository**.

- [ ]  Check the archival deposit (e.g., in Zenodo) to ensure it has the correct metadata. This includes the title (should match the paper title exactly) and author list (should match the people and the order in which they are listed on your paper exactly).

- [ ]  Please respond with the DOI of the archived version here along with the version number released.

As an aside, this package seems really well done and I'll likely use it in my own workflows!
