+++ paper/paper.bib
@@ -16,7 +16,7 @@
 @article{NaiveDifferencesOfExecutableCode,
     title =        {Naive differences of executable code},
     author =       {Percival, Colin},
-    journal =      {http://www.daemonology.net/bsdiff},
+    journal =      {https://www.daemonology.net/bsdiff/},
     year =         {2003},
     publisher =    {Citeseer},
 }
