+++ paper/paper.bib
@@ -16,7 +16,7 @@
 @article{NaiveDifferencesOfExecutableCode,
     title =        {Naive differences of executable code},
     author =       {Percival, Colin},
-    journal =      {https://www.daemonology.net/bsdiff/},
+    url =          {https://www.daemonology.net/bsdiff/},
     year =         {2003},
     publisher =    {Citeseer},
 }
