### 2025-06-02 f6b8fd1
Bump version to 0.1.8
