+++ keepdelta/__init__.py
@@ -1,1 +1,2 @@
+__version__ = "0.1.8"
 from .main import create, apply
