+++ keepdelta/__init__.py
@@ -1,2 +1,2 @@
-__version__ = "0.1.8"
+__version__ = "0.1.9"
 from .main import create, apply
