+++ pyproject.toml
@@ -13,7 +13,7 @@ keepdelta = ["py.typed"]
 
 [project]
 name = "keepdelta"
-version = "0.1.8"
+version = "0.1.9"
 description = "A Python Library for Human-Readable Data Differencing"
 readme = "README.md"
 license = {text = "MIT"}
